Crate substrate_telemetry

Telemetry utils.

telemetry macro may be used anywhere in the Substrate codebase in order to send real-time logging information to the telemetry server (if there is one). We use the async drain adapter of slog so that the logging thread doesn't get held up at all.

Re-exports

pub use slog;

Macros

telemetry
Translates to slog_scope::info, but contains an additional verbosity parameter which the log record is tagged with. Additionally the verbosity parameter is added to the record as a key-value pair.

Structs

Multiply
Multiply logging to all drains. This is similar to slog::Duplicate, which is limited to two drains though and doesn't support dynamic nesting at runtime.

TelemetryConfig
Configuration for telemetry.

TelemetryEndpoints

Constants

CONSENSUS_DEBUG

CONSENSUS_INFO

CONSENSUS_TRACE

CONSENSUS_WARN

SUBSTRATE_DEBUG
Log levels.

SUBSTRATE_INFO

Functions

init_telemetry
Initialize telemetry.

with_logger
Access the Logger for the current logging scope

Type Definitions

Telemetry
Telemetry service guard.